+++ torchdata/datapipes/__init__.py
@@ -1,6 +1,7 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
+from torch.utils.data import functional_datapipe
 from . import iter
 from . import map
 from . import utils
 
-__all__ = ["iter", "map", "utils"]
+__all__ = ["functional_datapipe", "iter", "map", "utils"]
