+++ torchdata/datapipes/iter/util/cacheholder.py
@@ -5,7 +5,8 @@ from collections import deque
 from os import path
 from typing import Deque, Optional
 
-from torchdata.datapipes.iter import IterDataPipe, FileLoader, functional_datapipe
+from torchdata.datapipes import functional_datapipe
+from torchdata.datapipes.iter import IterDataPipe, FileLoader
 from torchdata.datapipes.utils.common import _default_filepath_fn
