+++ torchdata/datapipes/iter/util/csvparser.py
@@ -1,7 +1,8 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 import csv
 
-from torchdata.datapipes.iter import IterDataPipe, functional_datapipe
+from torchdata.datapipes import functional_datapipe
+from torchdata.datapipes.iter import IterDataPipe
 
 
 class _CSVBaseParserIterDataPipe(IterDataPipe):
