+++ torchdata/datapipes/iter/util/hashchecker.py
@@ -1,7 +1,8 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 import hashlib
 
-from torchdata.datapipes.iter import IterDataPipe, functional_datapipe
+from torchdata.datapipes import functional_datapipe
+from torchdata.datapipes.iter import IterDataPipe
 
 
 @functional_datapipe("check_hash")
