+++ torchdata/datapipes/iter/util/ziparchivereader.py
@@ -7,7 +7,8 @@ from io import BufferedIOBase
 from typing import IO, Iterable, Iterator, Tuple, cast
 
 from torchdata.datapipes.utils.common import validate_pathname_binary_tuple
-from torchdata.datapipes.iter import IterDataPipe, functional_datapipe
+from torchdata.datapipes import functional_datapipe
+from torchdata.datapipes.iter import IterDataPipe
 
 # TODO(VitalyFedyunin): This file copy-pasted from the pytorch repo
 # nuke source class when repo is open-sourced
